refactor(client): split ws_message_handler into feature part files (#755)
